Completed Order preliminary with no outbox.

+++ order/src/main/java/com/evolutionnext/order/domain/aggregate/order/OrderItem.java
@@ -3,5 +3,7 @@ package com.evolutionnext.order.domain.aggregate.order;
 
 import com.evolutionnext.order.domain.aggregate.product.ProductId;
 
-public record OrderItem(ProductId productId, int quantity, int price) {
+import java.math.BigDecimal;
+
+public record OrderItem(ProductId productId, int quantity, BigDecimal price) {
 }
