+++ order/src/main/java/com/evolutionnext/order/domain/aggregate/product/Product.java
@@ -1,5 +1,7 @@
 package com.evolutionnext.order.domain.aggregate.product;
 
 
-public record Product(ProductId productId, String name, int price) {
+import java.math.BigDecimal;
+
+public record Product(ProductId productId, String name, BigDecimal price) {
 }
